+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.0.6]
+
+### Changed
+
+- Using a Name Tag with a Custom Name on a Villager now fully resets the Villager's profession in their name.
+- Changed Minecraft version requirement to allow the mod to run on versions past 1.21.1.
+
 ## [1.0.5]
 
 ### Fixed